requirements/base.txt
Django==1.6.2
django-guardian==1.2.0
django-mptt==0.6.0
django-mptt-admin==0.1.8
django-suit==0.2.7
South==0.8.4
Sphinx==1.2.2
Unipath==1.0

